Removing some double-to-something warnings.
git-svn-id: http://svn.seqan.de/seqan/trunk@7889 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/apps/mason/simulate_sanger.h
@@ -334,7 +334,7 @@ void buildSimulationInstructions(ReadSimulationInstruction<SangerReads> & inst, 
                 stdDev = options.qualityErrorStartStdDev + pos * (options.qualityErrorEndStdDev - options.qualityErrorStartStdDev);
             }
             PDF<Normal> pdf(mean, stdDev);
-            appendValue(inst.qualities, pickRandomNumber(rng, pdf));
+            appendValue(inst.qualities, static_cast<int>(pickRandomNumber(rng, pdf)));
         }
     }
 }
